deleted BigInteger class

+++ psifos/crypto/sharedpoint.py
@@ -5,32 +5,30 @@ SharedPoint classes for Psifos.
 """
 
 from psifos.serialization import SerializableList, SerializableObject
-from psifos.crypto.utils import BigInteger
-
 
 class Signature(SerializableObject):
     def __init__(self, challenge, response) -> None:
-        self.challenge: BigInteger = BigInteger(challenge)
-        self.response: BigInteger = BigInteger(response)
+        self.challenge: int = int(challenge)
+        self.response: int = int(response)
 
 
 class Certificate(SerializableObject):
     def __init__(self, signature_key, encryption_key, signature) -> None:
-        self.signature_key: BigInteger = BigInteger(signature_key)
-        self.encryption_key: BigInteger = BigInteger(encryption_key)
+        self.signature_key: int = int(signature_key)
+        self.encryption_key: int = int(encryption_key)
         self.signature: Signature = Signature(**signature)
 
 
 class Coefficient(SerializableObject):
     def __init__(self, coefficient, signature) -> None:
-        self.coefficient: BigInteger = BigInteger(coefficient)
+        self.coefficient: int = int(coefficient)
         self.signature: Signature = Signature(**signature)
 
 
 class Point(SerializableObject):
     def __init__(self, alpha, beta, signature) -> None:
-        self.alpha: BigInteger = BigInteger(alpha)
-        self.beta: BigInteger = beta
+        self.alpha: int = int(alpha)
+        self.beta: int = beta
         self.signature: Signature = Signature(**signature)
